feat: add next/fonts to Storybook
